requirements.txt
einops
imgcat
requests
pytest
click
joblib
tqdm
ipykernel
notebook
ipywidgets
jupyter
flask
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
black
wandb
ftfy
tensorboard
pygame
plotly
sympy
accelerate
pre-commit
diffusers
gdown
jaxtyping
typeguard<4
streamlit
torchinfo
torchvision
frozendict
streamlit-chat
pytorch-lightning
openai
transformer_lens
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
